Update calendar.js
Modified _loadEvents to support async feed requests. Also, after the async call is complete, it uses the contents of the, optional, schema property to map the data returned to the default Event schema. The Schema is an object with a property named for each property in the Event schema that you want to map.  Unmapped properties are left unaltered. The value of a schema property can be either a String or Function. If it is a string then the value is the name of the inbound field.  If it is a function then it is passed a reference to an and Event object; the function must return a String, Date or Number.